+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Digestiflow CLI Client Changelog
 
+## HEAD (unreleased)
+
+- Registring indices seen in >0.1% of all reads by default now.
+- Allowing to configure minimal fraction of reads to show an index for index to turn up in histogram.
+
 ## v0.1.1
 
 - Adding lincense.